Initial attempt to improve preview functionality
This commit attempts to leverage {s,g}etDomainUnstructuredSkipMeshing to
improve the preview functionality that I am trying to implement. I do
not love this as I think it adds quite a bit of complexity (perhaps more
than it is worth) but it does seem to be a step closer to how I
envisioned this working in the first place. This is not thoroughly
tested and it has been hastily committed; proceed with caution.